MultiCropDiseaseDetector/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
click==8.3.0
dnspython==2.8.0
efficientnet_pytorch==0.7.1
email-validator==2.3.0
fastapi==0.117.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.1
filelock==3.13.1
fsspec==2024.6.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
packaging==25.0
pillow==11.0.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
sentry-sdk==2.39.0
setuptools==70.2.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.48.0
sympy==1.13.3
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
